refactor: 책 상세페이지, 마이페이지에서 userNo파라미터로 받아오는 것으로 수정, 책 상세페이지 동일 저자 책 css 수정, 마이페이지 이미지 css수정, 로그인 후 index.jsp 데이터 요청 파라미터로 리턴값 수정

+++ src/main/java/com/leafpage/controller/book/DetailPageController.java
@@ -18,7 +18,7 @@ public class DetailPageController implements Controller {
     public String handleRequest(HttpServletRequest request, HttpServletResponse response) throws IOException {
 
         HttpSession session = request.getSession();
-        long userNo = (long) session.getAttribute("userNo");
+        Long userNo = ((Integer) session.getAttribute("userNo")).longValue();
         System.out.println("userNo"+userNo);
         String isbn = request.getParameter("isbn");
